Test cases updated

+++ src/main/java/com/codedecode/foodcatalogue/dto/FoodItemDTO.java
@@ -5,6 +5,6 @@ public record FoodItemDTO(Long id,
                           String itemDescription,
                           Boolean isVeg,
                           Long price,
-                          Integer restaurantId,
+                          Long restaurantId,
                           Integer quantity) {
 }
